+++ changelog.md
@@ -4,8 +4,10 @@ All notable changes to `laravel-alma-client` will be documented in this file.
 
 ## **Version 1.0.1**
 
-### Changed
+### Fixed
+- Instanciate Alma Client take good config key
 
+### Changed
 - Require Alma client ~1 instead of fixed ^1.8
 
 ## **Version 1.0.0**